Changed code to use ImageRegistry and registering images
The code now uses ImageRegistry with the before images in beforeImageRegistry, after images in afterImageRegistry and minimaps in minimapRegistry.

Compared to the previous commit a lot of images has now been registered in them. Each of the different registers has their own fallback.

It should be mentioned that hungry groups show a different fallback image compared to rooms without a group. This is not a bug.

An idea is to add a class called RegistryInitialiser where the registers get initialised so a lot of the registering of images can get moved out of the Controller and therefore out of the logic.